requirements.txt
pywin32
opencv-python<=4.5.1.48
dhash
sewar
Pillow
numpy
tk
requests2
matplotlib
scipy
scikit-learn==0.24.0
readerwriterlock
scrcpy-client
psutil
PySide6
flask
flask-socketio